requirements.txt
# Message
Flask==2.1.1
google-cloud-pubsub